# Log SDK optimization - Improve perf 15%-30%