--- a/opentelemetry-sdk/benches/log_processor.rs
+++ b/opentelemetry-sdk/benches/log_processor.rs
@@ -19,6 +19,7 @@ use std::{
 
 use criterion::{criterion_group, criterion_main, Criterion};
 use opentelemetry::logs::{LogRecord as _, LogResult, Logger as _, LoggerProvider as _, Severity};
+use opentelemetry::InstrumentationLibrary;
 use opentelemetry_sdk::{
     logs::LogData,
     logs::{LogProcessor, LogRecord, Logger, LoggerProvider},
@@ -45,7 +46,7 @@ fn create_log_record(logger: &Logger) -> LogRecord {
 struct NoopProcessor;
 
 impl LogProcessor for NoopProcessor {
-    fn emit(&self, _data: &mut LogData) {}
+    fn emit(&self, _data: &mut LogRecord, _library: &InstrumentationLibrary) {}
 
     fn force_flush(&self) -> LogResult<()> {
         Ok(())
@@ -60,7 +61,7 @@ impl LogProcessor for NoopProcessor {
 struct CloningProcessor;
 
 impl LogProcessor for CloningProcessor {
-    fn emit(&self, data: &mut LogData) {
+    fn emit(&self, data: &mut LogRecord, _library: &InstrumentationLibrary) {
         let _data_cloned = data.clone();
     }
 
@@ -75,8 +76,8 @@ impl LogProcessor for CloningProcessor {
 
 #[derive(Debug)]
 struct SendToChannelProcessor {
-    sender: std::sync::mpsc::Sender<LogData>,
-    receiver: Arc<Mutex<std::sync::mpsc::Receiver<LogData>>>,
+    sender: std::sync::mpsc::Sender<(LogRecord, InstrumentationLibrary)>,
+    receiver: Arc<Mutex<std::sync::mpsc::Receiver<(LogRecord, InstrumentationLibrary)>>>,
 }
 
 impl SendToChannelProcessor {
@@ -103,9 +104,8 @@ impl SendToChannelProcessor {
 }
 
 impl LogProcessor for SendToChannelProcessor {
-    fn emit(&self, data: &mut LogData) {
-        let data_cloned = data.clone();
-        let res = self.sender.send(data_cloned);
+    fn emit(&self, record: &mut LogRecord, library: &InstrumentationLibrary) {
+        let res = self.sender.send((record.clone(), library.clone()));
         if res.is_err() {
             println!("Error sending log data to channel {0}", res.err().unwrap());
         }