--- a/opentelemetry-sdk/benches/log_processor.rs
+++ b/opentelemetry-sdk/benches/log_processor.rs
@@ -20,10 +20,7 @@ use std::{
 use criterion::{criterion_group, criterion_main, Criterion};
 use opentelemetry::logs::{LogRecord as _, LogResult, Logger as _, LoggerProvider as _, Severity};
 use opentelemetry::InstrumentationLibrary;
-use opentelemetry_sdk::{
-    logs::LogData,
-    logs::{LogProcessor, LogRecord, Logger, LoggerProvider},
-};
+use opentelemetry_sdk::logs::{LogProcessor, LogRecord, Logger, LoggerProvider};
 
 // Run this benchmark with:
 // cargo bench --bench log_processor