--- a/opentelemetry-sdk/src/logs/log_emitter.rs
+++ b/opentelemetry-sdk/src/logs/log_emitter.rs
@@ -1,5 +1,5 @@
 use super::{BatchLogProcessor, LogProcessor, LogRecord, SimpleLogProcessor, TraceContext};
-use crate::{export::logs::LogExporter, logs::LogData, runtime::RuntimeChannel, Resource};
+use crate::{export::logs::LogExporter, runtime::RuntimeChannel, Resource};
 use opentelemetry::{
     global,
     logs::{LogError, LogResult},
@@ -254,28 +254,24 @@ impl opentelemetry::logs::Logger for Logger {
     }
 
     /// Emit a `LogRecord`.
-    fn emit(&self, record: Self::LogRecord) {
+    fn emit(&self, mut record: Self::LogRecord) {
         let provider = self.provider();
         let processors = provider.log_processors();
         let trace_context = Context::map_current(|cx| {
             cx.has_active_span()
                 .then(|| TraceContext::from(cx.span().span_context()))
         });
-        let mut log_record = record;
-        if let Some(ref trace_context) = trace_context {
-            log_record.trace_context = Some(trace_context.clone());
-        }
-        if log_record.observed_timestamp.is_none() {
-            log_record.observed_timestamp = Some(SystemTime::now());
-        }
 
-        let mut data = LogData {
-            record: log_record,
-            instrumentation: self.instrumentation_library().clone(),
-        };
+        //let mut log_record = record;
+        if let Some(ref trace_context) = trace_context {
+            record.trace_context = Some(trace_context.clone());
+        }
+        if record.observed_timestamp.is_none() {
+            record.observed_timestamp = Some(SystemTime::now());
+        }
 
         for p in processors {
-            p.emit(&mut data);
+            p.emit(&mut record, self.instrumentation_library());
         }
     }
 
@@ -332,7 +328,7 @@ mod tests {
     }
 
     impl LogProcessor for ShutdownTestLogProcessor {
-        fn emit(&self, _data: &mut LogData) {
+        fn emit(&self, _data: &mut LogRecord, _library: &InstrumentationLibrary) {
             self.is_shutdown
                 .lock()
                 .map(|is_shutdown| {
@@ -562,7 +558,7 @@ mod tests {
     }
 
     impl LogProcessor for LazyLogProcessor {
-        fn emit(&self, _data: &mut LogData) {
+        fn emit(&self, _data: &mut LogRecord, _library: &InstrumentationLibrary) {
             // nothing to do.
         }