--- a/stress/src/logs.rs
+++ b/stress/src/logs.rs
@@ -9,6 +9,7 @@
     ~38 M /sec
 */
 
+use opentelemetry::InstrumentationLibrary;
 use opentelemetry_appender_tracing::layer;
 use opentelemetry_sdk::logs::{LogProcessor, LoggerProvider};
 use tracing::error;
@@ -20,7 +21,12 @@ mod throughput;
 pub struct NoOpLogProcessor;
 
 impl LogProcessor for NoOpLogProcessor {
-    fn emit(&self, _data: &mut opentelemetry_sdk::logs::LogData) {}
+    fn emit(
+        &self,
+        _record: &mut opentelemetry_sdk::logs::LogRecord,
+        _library: &InstrumentationLibrary,
+    ) {
+    }
 
     fn force_flush(&self) -> opentelemetry::logs::LogResult<()> {
         Ok(())