## Changes

Optimize Log SDK by avoiding unnecessary copy operation and redirections 

- `LogData` structure is removed.
- `Logger::emit()` method is modified for unnecessary copy/assignment operation.
- `LogProcessor::emit` method is modified to directly take reference of `LogRecord` and `InstrumentationLibrary` instead of `LogData`. This simplifies the API, and more importantly ones less level of redirection while accessing these components.

Existing:
```rust
fn emit(&self, _: &mut LogData)
```

Updated:
```rust
fn emit(&self, record: &mut LogRecord, library: &InstrumentationLibrary);
```

**Benchmarks and stress:** (courtesy chatgpt to generate table from the results):

| **Benchmark**                                | **Main Branch (ns)**                         | **PR Branch (ns)**                            | **Change**                      | **Performance**           |
|----------------------------------------------|----------------------------------------------|----------------------------------------------|---------------------------------|---------------------------|
| `ot_layer_enabled`                           | 264.70 - 268.47                              | 216.33 - 216.69                              | -17.979% to -18.548%            | Improved                   |
| `Logging_Comparable_To_Appender`             | 134.11 - 134.30                              | 96.087 - 96.240                              | -28.174% to -28.378%            | Improved                   |
| `log_noop_processor`                         | 141.65 - 141.81                              | 102.56 - 102.70                              | -27.468% to -27.635%            | Improved                   |
| `log_cloning_processor`                      | 238.92 - 239.35                              | 178.97 - 179.24                              | -25.048% to -25.337%            | Improved                   |
| `log_clone_and_send_to_channel_processor`    | 620.69 - 628.54                              | 582.68 - 584.69                              | -6.0463% to -6.9739%            | Improved                   |
| `LogExporterWithFuture`                      | 284.68 - 285.04                              | 129.92 - 130.31                              | -54.197% to -54.449%            | Improved                   |
| `LogExporterWithoutFuture`                   | 250.91 - 251.26                              | 98.156 - 98.299                              | -60.799% to -60.891%            | Improved                   |

| **Stress (Throughput)**                      | **Main Branch (iterations/sec)**             | **PR Branch (iterations/sec)**               | **Change**                      | **Performance**           |
|----------------------------------------------|----------------------------------------------|----------------------------------------------|---------------------------------|---------------------------|
| Average                                      | 38,812,133                                   | 44,491,457                                   | +14.7%                          | Improved                   |
 

## Merge requirement checklist

* [x] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [x] Unit tests added/updated (if applicable)
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [x] Changes in public API reviewed (if applicable)